More refactoring required to fully implement the dual prompts
Major refactoring to simplify the MFLUX code base required while adding dual prompt support to all generate commands, image generation and callbacks. Identified many areas of the code that could be simplifies with the removal of much of the unnecessary duplicated code. All new code is backward compatible to avoid breaking existing projects that are using this code. The code is not yet fully refactored and will not run at this stage.